+++ src/repository/mapper/manual_transaction.go
@@ -10,12 +10,12 @@ type ManualTransaction struct {
 	//Dados de transações
 	ID        int       `gorm:"column:id"`
 	Symbol    string    `gorm:"column:symbol"`
-	Quantity  int       `gorm:"column:quantity"`
+	Quantity  float64   `gorm:"column:quantity"`
 	Price     float64   `gorm:"column:price"`
 	TradeDate time.Time `gorm:"column:trade_date"`
 
 	//Dados calculados com base nos eventos corporativos
-	PostEventQuantity int     `gorm:"column:post_event_quantity"`
+	PostEventQuantity float64 `gorm:"column:post_event_quantity"`
 	PostEventPrice    float64 `gorm:"column:post_event_price"`
 
 	// Dados do evento corporativo que estão na Financial
